requirements.txt
apifairy==1.3.0
apispec==6.3.0
binary-reader==1.4.3
blinker==1.6.2
certifi==2023.7.22
cffi==2.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
elastic-transport==8.4.0
elasticsearch==8.9.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
flask-marshmallow==0.15.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==1.26.16
webargs==8.3.0
Werkzeug==2.3.7
zipp==3.16.2